fix: create element block not working
